revamp language switcher to hopefully work more reliably
* this actually can reset to the "System Default"
* switches immediately in Settings and Onboarding
* switches MainActivity after a tab change
